refac: changed Initialization of Authorization logic
