+++ src/zbw.Auftragsverwaltung.Core/Common/Contracts/IRepository.cs
@@ -7,9 +7,9 @@ using zbw.Auftragsverwaltung.Core.Common.Interfaces;
 
 namespace zbw.Auftragsverwaltung.Core.Common.Contracts
 {
-    public interface IRepository<TI> where TI : class, IEntity
+    public interface IRepository<TI, TKey> where TI : class
     {
-        Task<TI> GetByIdAsync(Guid id);
+        Task<TI> GetByIdAsync(TKey id);
         Task<IReadOnlyList<TI>> ListAsync();
         Task<IReadOnlyList<TI>> ListAsync(Expression<Func<TI, bool>> predicate);
         Task<TI> AddAsync(TI entity);
