+++ src/zbw.Auftragsverwaltung.Core/Customers/BLL/CustomerBll.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using zbw.Auftragsverwaltung.Core.Common.Contracts;
 using zbw.Auftragsverwaltung.Core.Common.DTO;
+using zbw.Auftragsverwaltung.Core.Customers.Contracts;
 using zbw.Auftragsverwaltung.Core.Customers.Dto;
 using zbw.Auftragsverwaltung.Core.Customers.Entities;
 using zbw.Auftragsverwaltung.Core.Customers.Interfaces;
@@ -15,10 +16,10 @@ namespace zbw.Auftragsverwaltung.Core.Customers.BLL
 {
     public class CustomerBll : ICustomerBll
     {
-        private readonly IRepository<Customer> _customerRepository;
+        private readonly ICustomerRepository _customerRepository;
         private readonly IMapper _mapper;
 
-        public CustomerBll(IRepository<Customer> customerRepository, IMapper mapper)
+        public CustomerBll(ICustomerRepository customerRepository, IMapper mapper)
         {
             _customerRepository = customerRepository;
             _mapper = mapper;
