+++ src/zbw.Auftragsverwaltung.Core/Customers/Contracts/ICustomerRepository.cs
@@ -1,9 +1,10 @@
-﻿using zbw.Auftragsverwaltung.Core.Common.Contracts;
+﻿using System;
+using zbw.Auftragsverwaltung.Core.Common.Contracts;
 using zbw.Auftragsverwaltung.Core.Customers.Entities;
 
 namespace zbw.Auftragsverwaltung.Core.Customers.Contracts
 {
-    public interface ICustomerRepository : IRepository<Customer>
+    public interface ICustomerRepository : IRepository<Customer, Guid>
     {
 
     }
