+++ src/zbw.Auftragsverwaltung.Core/Users/Interfaces/IUserBll.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using zbw.Auftragsverwaltung.Core.Common.Interfaces;
 using zbw.Auftragsverwaltung.Core.Users.Dto;
 using zbw.Auftragsverwaltung.Core.Users.Entities;
@@ -9,5 +11,10 @@ namespace zbw.Auftragsverwaltung.Core.Users.Interfaces
 {
     public interface IUserBll : ICrudBll<UserDto, User, Guid>
     {
+        public Task<IdentityResult> AddToRole(User user, string role);
+
+        public Task<IdentityResult> RemoveFromRole(User user, string role);
+
+        public Task<IdentityResult> ChangeUserPassword(User user, string currentPassword, string newPassword);
     }
 }
