+++ src/zbw.Auftragsverwaltung.Infrastructure/Customers/DAL/CustomerRepository.cs
@@ -9,7 +9,7 @@ using zbw.Auftragsverwaltung.Infrastructure.Common.Repositories;
 
 namespace zbw.Auftragsverwaltung.Infrastructure.Customers.DAL
 {
-    public class CustomerRepository : BaseRepository<Customer>, ICustomerRepository
+    public class CustomerRepository : BaseRepository<Customer, Guid, OrderManagementContext>, ICustomerRepository
     {
         public CustomerRepository(OrderManagementContext dbContext) : base(dbContext)
         {
