+++ src/zbw.Auftragsverwaltung.Infrastructure/zbw.Auftragsverwaltung.Infrastructure.csproj
@@ -2,18 +2,17 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
+    <UserSecretsId>0dac7593-1a87-478b-a930-09f32eaddc39</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
+    <Folder Include="Common\Configurations\" />
     <Folder Include="Common\Services\" />
     <Folder Include="Customers\Services\" />
     <Folder Include="Users\Services\" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="3.1.10" />
-    <PackageReference Include="Microsoft.AspNetCore.Authorization.Policy" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.13" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.7" />
